- tauri.conf.json: AxInvest + v2.0.0
- locale JSON: 上游重建 + AxInvest stockAnalysis 键还原
- i18n allowlist: 采用上游 269 条目
- gateway: Rust 2024 edition 兼容
- SubAgentCard/SettingsSidebar: 合并双方改动

+++ src-tauri/crates/core/src/repo/gateway_request_log.rs
@@ -4,6 +4,7 @@ use crate::types::GatewayRequestLog;
 use crate::utils::{gen_id, now_ts};
 use sea_orm::*;
 
+#[allow(clippy::too_many_arguments)]
 pub async fn record_request_log(
     db: &DatabaseConnection,
     key_id: &str,
@@ -14,8 +15,8 @@ pub async fn record_request_log(
     provider_id: Option<&str>,
     status_code: i32,
     duration_ms: i32,
-    request_tokens: i32,
-    response_tokens: i32,
+    request_tokens: i64,
+    response_tokens: i64,
     error_message: Option<&str>,
 ) -> Result<()> {
     gateway_request_logs::ActiveModel {
